Adding main.cpp file to rinremoval-recon-polar
This is the code that actually handles the ring removal and file I/O.
File I/O is done through libtiff, so this code must be compiled
by linking to the local tiff library. This version works (the executable
can be called with no parameters to see proper usage), but it takes
a significant time, because the median filter just utilizes a bare-bones
quick sort algorithm.  It includes the skeleton of a few functions that
will hopefully work faster, but nothing actually working as of this commit

It also has two functions to do the mean filtering, a bare bones function
and a "rolling" function that's slightly faster, but introduces a small floating
point rounding error (not significant for images I'm dealing with, so it
uses the faster one right now). It assumes that the images passed to it are
32 bit floating point tifs.