Starting to move code to seperate classes
To make it easier to implement multiple ring removal algorithms in
the future, I'm starting to seperate out the recon-polar code into
several seperate classes: and I/O class (TiffIO), a image filter
class (ImageFilterClass), and an image transform class
(ImageTransformClass).  This way, if I see another ring removal
algorithm I'd like to try to implement, I can just load the necessary
files and not worry about re-coding certain functions.

N.B. - Is it possible to make a broad enough tiff reader to be used
with gridRec as well? Is that necessary, or is the GridReadWrite class
unique enough that it probably won't be applicable? Something to think
about.