Finished factoring out code, compiles and runs, but has major issue
The code is now pretty much fully factored out.  I changed how the
inverse polar transformation was behaving, as it wasn't working
correctly when the center passed to it wasn't near the center of the
image.  Now it should handle all centers correctly (though I should note
that right now the filter only works on acircular area centered at
the center provided - this means it finds the largest circle it can that
doesn't extend past any sides of the image, and only filters that area).

However, though the code compiles and runs, something is wrong with the
way that it's handling the filtering - the ring filter no longer finds and
removes rings.  The next step is going in and seeing what could have changed
that would cause this weird behaviour (I have a feeling it occured when I
changed teh radial filter from returning a single value to a void function
that works on an arbitrary area of the image...)

One possible area to improve in the future is to have it filter the entire
image? the problem is that I'm not quite sure how to fill the polar space
for (r, theta) pairs that don't exist in image space, and that would
effect the filtering behaviour...