Finished factoring code, fixed all issues, works the same as integrated code
Figure out where the bug was (it was actually in the mean filter - the sum
variable wasn't being reset between columns, so the mean was continually
growing in each of the 3 regions).  Now that that's fixed, it works just like
the integrated version did (and is actually a bit faster, as I cleaned up/
changed some of the calls to make them a bit more concise).

First things to do are go through and see if there's anywhere else I can clean
up the code (possibly in polar transform, i think the memory allocaiton is
different there/uses a lot of allocations, instead of doing the recursive setup
like the other functions).  Next, it's to start looking at the median code to
see if there's still anything that I'm doing that's wrong/causing it to take a
long time.  If it looks good/I can't see anything, then I'll start looking at
additional algorithms to try to implement (but at least I know that I have
a separate function that will work for arbitrary image subdomains).