Starting to add a faster median filter
Starting to add a new function for doing a faster median filter.
Right now, this function will basically be a "rolling" median filter
that will only work with a filter width and ring width of the same
size, so that no pixel is sampled more than once.  I need to run some
tests to see what sort of results I can get because of this (try running
the slow version with various size kernels, but leave the ring width
the same to see how oversampling influences the median... as far as
I can tell it really shouldn't be that noticable, but need to test it
out...)

The way I see this function working is: use quicksort for first element of
the row, but in sorting, also swap elements of an integer array that
tracks the "position" of each element.  Then, using the position array,
remove the last element, and insert the new element into the correct
position, swapping elements along the way.  This way, once the next
element is added to the correct position, the array will still be in order
and the median can be immediately returned.