Finished the "x-axis" portion of fast median filter
This fucntion does a quick sort on an array of values while also ordering
an array of positions.  Then it does a recursive "bubble" to place a new
value in the correct position in the array.  Need to see if this compiles,
and if so, how fast it runs/if it gives the same results

Note: when checking the results, need to check it against the "slow"
fucntion that uses the same size kernel/ring width, because right now the
fast version doesn't support oversampling for the median array.