Changed bilinear interpolation for inverse polar transform
Realized that the image create by performing a polar transform
immediately followed by the inverse transform can actually differ
from the image by as much as 10-15%, which is a problem.  Looking
at the code, I found that the inverse transform wasn't properly
interpolating (I was treating the polar points as a regular grid,
but it's not, the area between 4 points can change quite a bit
depending on the radius).

However, even after changing this, the tranformed image is still
differing from the original by ~10% (nearly 0% near angles 0, 90,
180, and 270, but bad near 45, 135, 225, and 315). I have a feeling
this is because there are many points that fall within a single x,y
grid point, but just choosing the four for the interpolation is
causing the data to become "blurred".  Not sure how to fix this at
the moment, but need to continue looking at it...