Changed image tranforms to find better settings
Polar/Inverse transform were really blurring the initial data.  After
fixing a few steps in the interpolation process, I changed the way that
the polar coordinates are scaled, so that a step in the radius direction
is 1/5 the unit square in cartesian coordinates.  This greatly improves
the blurring effect at boundaries, but because the image is larger, it
also increases computational time.  Think about adding this as a parameter?
It could be nice to fiddle with, but the user isn't really going to know
what it does... (need to focus on making all parameters as transparent/even
dumb them down a little so that it's not confusing for the end user).