Added two scaling parameters to the transform functions
These factors allow for greater control of how much oversampling
is done in the radial and azimuthal directions.  It also allows
for the ring_width the user enters to be adjusted accordingly
(so that the number of oversampled pixels that make up the filter's
kernel width corresponds to the desired number of image pixels)

Also introduced (but haven't implemented) an overhang feature for
the polar and inverse polar transform, so that filters can run without
worrying about edge effects (the edges will be built in to the polar
image).