Fixed bug in polar transform, added more messages when verbose
The polar transform had a bug in it where it would try to access
pixels that were out of the bounds of the original image.  This
code should have that fixed.

I also put in another verbose message for when a file is being
open, so it's easier to see what stage the program is at.