1 pixel ring issue mostly fixed
Figured out the reason for the one pixel rings remaining was due to
the way I was oversampling the image in the r direction.  After
playing around a little bit, I was able to fix it for the most part.
However, the undersampled image is introducing a bit of noise around
the center of the image.  I feel like there must be a way to oversample
and still filter correctly, to remove the rings and minimize noise.
Still need to play around with it a little.